Fixes https://github.com/gitpod-io/gitpod/issues/14714
